# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Secure Gift Card Manager API
  version: 0.1.0
  description: >
    Initial API contract for the local MVP. The implementation must keep this
    file in sync with backend routes and QA requirement IDs.
servers:
  - url: http://localhost:3001
    description: Local API server
...
paths:
...
  /api/admin/support-policy:
    get:
      tags: [Admin]
      summary: Get support access policy
      operationId: getSupportPolicy
      responses:
        "200":
          description: Support access policy
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ObjectResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SupportPolicy"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [Admin]
      summary: Update support access policy
      operationId: updateSupportPolicy
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportPolicyUpdateRequest"
      responses:
        "200":
          description: Updated support access policy
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ObjectResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SupportPolicy"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/data-policy:
    get:
      tags: [Admin]
      summary: Get data retention policy
      operationId: getDataPolicy
      responses:
        "200":
          description: Data retention policy
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ObjectResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DataPolicy"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [Admin]
      summary: Update data retention policy
      operationId: updateDataPolicy
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataPolicyUpdateRequest"
      responses:
        "200":
          description: Updated data retention policy
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ObjectResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DataPolicy"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/retention/run:
    post:
      tags: [Admin]
      summary: Preview or run data retention deletion
      operationId: runDataRetention
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetentionRunRequest"
      responses:
        "200":
          description: Retention result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetentionRunResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/data-export:
    post:
      tags: [Admin]
      summary: Export sanitized account data without card credentials or key material
      operationId: exportSanitizedAccountData
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SanitizedDataExportRequest"
      responses:
        "200":
          description: Sanitized export
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ObjectResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SanitizedAccountExport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/data-delete:
    post:
      tags: [Admin]
      summary: Delete account inventory data while preserving users and audit trail
      operationId: deleteAccountInventoryData
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteInventoryDataRequest"
      responses:
        "200":
          description: Inventory deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteInventoryDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 128
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 16
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Not authenticated or locked
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated request failed CSRF, origin, or authorization checks
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    User:
      type: object
      required: [id, accountId, displayName, role, createdAt, updatedAt]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        email:
          type: string
          nullable: true
        displayName:
          type: string
        role:
          type: string
          enum: [owner, admin, operator, viewer]
        disabledAt:
          type: string
          format: date-time
          nullable: true
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SupportPolicy:
      type: object
      required:
        - supportAccessEnabled
        - supportContact
        - supportPolicyUrl
        - supportNotes
        - supportUpdatedAt
        - supportUpdatedByUserId
      properties:
        supportAccessEnabled:
          type: boolean
        supportContact:
          type: string
        supportPolicyUrl:
          type: string
        supportNotes:
          type: string
        supportUpdatedAt:
          type: string
          format: date-time
          nullable: true
        supportUpdatedByUserId:
          type: integer
          nullable: true
    SupportPolicyUpdateRequest:
      type: object
      required: [unlockSecret, supportAccessEnabled]
      properties:
        unlockSecret:
          type: string
        supportAccessEnabled:
          type: boolean
        supportContact:
          type: string
          maxLength: 200
        supportPolicyUrl:
          type: string
          maxLength: 500
        supportNotes:
          type: string
          maxLength: 1000
    DataPolicy:
      type: object
      required:
        - auditRetentionDays
        - idempotencyRetentionDays
        - sessionRetentionDays
        - loginAttemptRetentionDays
      properties:
        auditRetentionDays:
          type: integer
          minimum: 1
          maximum: 3650
        idempotencyRetentionDays:
          type: integer
          minimum: 1
          maximum: 3650
        sessionRetentionDays:
          type: integer
          minimum: 1
          maximum: 3650
        loginAttemptRetentionDays:
          type: integer
          minimum: 1
          maximum: 3650
    DataPolicyUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/DataPolicy"
        - type: object
          required: [unlockSecret]
          properties:
            unlockSecret:
              type: string
    RetentionRunRequest:
      type: object
      required: [unlockSecret]
      properties:
        unlockSecret:
          type: string
        dryRun:
          type: boolean
          default: false
        confirmation:
          type: string
          description: Required as PURGE when dryRun is false.
    RetentionCounts:
      type: object
      required: [auditLog, idempotencyKeys, webSessions, loginAttempts]
      properties:
        auditLog:
          type: integer
          minimum: 0
        idempotencyKeys:
          type: integer
          minimum: 0
        webSessions:
          type: integer
          minimum: 0
        loginAttempts:
          type: integer
          minimum: 0
    RetentionRunResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [dryRun, counts]
          properties:
            dryRun:
              type: boolean
            counts:
              $ref: "#/components/schemas/RetentionCounts"
    SanitizedDataExportRequest:
      type: object
      required: [unlockSecret, confirmation]
      properties:
        unlockSecret:
          type: string
        confirmation:
          type: string
          enum: [EXPORT]
    SanitizedAccountExport:
      type: object
      required: [schemaVersion, exportType, exportedAt, warning, counts]
      properties:
        schemaVersion:
          type: integer
          enum: [1]
        exportType:
          type: string
          enum: [sanitized_account_json]
        exportedAt:
          type: string
          format: date-time
        warning:
          type: string
        accounts:
          type: array
          items:
            type: object
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        appSettings:
          type: array
          items:
            type: object
        deals:
          type: array
          items:
            type: object
        cards:
          type: array
          items:
            type: object
            description: Sanitized card rows omit cardNumber, pin, billingZip, encrypted values, and blind indexes.
        transactions:
          type: array
          items:
            type: object
        usages:
          type: array
          items:
            type: object
        importJobs:
          type: array
          items:
            type: object
        auditLog:
          type: array
          items:
            type: object
        counts:
          type: object
          additionalProperties:
            type: integer
            minimum: 0
    DeleteInventoryDataRequest:
      type: object
      required: [unlockSecret, confirmation]
      properties:
        unlockSecret:
          type: string
        confirmation:
          type: string
          enum: [DELETE_ACCOUNT_DATA]
    DeleteInventoryDataResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [deleted, counts]
          properties:
            deleted:
              type: boolean
            counts:
              type: object
              additionalProperties:
                type: integer
                minimum: 0
...
    ObjectResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            fieldErrors:
              type: array
              items:
                type: object
                required: [field, code, message]
                properties:
                  field:
                    type: string
                  code:
                    type: string
                  message:
                    type: string